+++ tests/data/project3/file1.py
@@ -1,1 +1,1 @@
-# TIMEBOMB - FEWTURE-BOOM: report me
+# TIMEBOMB - FEWTURE-BOOM: report as an annotation